Use LocalFileLoader to check if the file exists. Also move GarchingArchiveModel to the loaders
